v6.0.0
- Fix non-standard link format from some RSS sources
- Update dependencies
