+++ package.json
@@ -1,5 +1,5 @@
 {
-  "version": "6.0.0rc4",
+  "version": "6.0.0",
   "name": "feed-reader",
   "description": "To read and normalize RSS/ATOM/JSON feed data",
   "homepage": "https://www.npmjs.com/package/feed-reader",
@@ -25,14 +25,14 @@
   },
   "dependencies": {
     "axios": "^0.27.2",
-    "bellajs": "^11.0.3",
+    "bellajs": "^11.0.4",
     "fast-xml-parser": "^4.0.9",
     "html-entities": "^2.3.3"
   },
   "devDependencies": {
-    "esbuild": "^0.14.49",
-    "jest": "^28.1.2",
-    "nock": "^13.2.8"
+    "esbuild": "^0.14.51",
+    "jest": "^28.1.3",
+    "nock": "^13.2.9"
   },
   "standard": {
     "ignore": [
